+++ extensions/dev-to/CHANGELOG.md
@@ -1,3 +1,14 @@
 # dev-to Changelog
 
-## [Initial Version] - 2022-08-15
+## [ Doc ] - 2022-08-18
+
+- Update README
+- Add Screenshots
+
+## [Initial Version] - 2022-08-17
+
+Initial version code
+
+- reading list
+- manage articles
+- create article